- Fixed aspect ratio miscalculation - MovingCircle's now start in random positions - Random number generation is now unique with each experiment - Regenerated all UI members with proper labels to make the code look prettier - Fixed Active Window tracking, this fixes the second window seeming to never render - Added active window focusing so that the ESCAPE key can be used once the experiment has started.

+++ RenderTargetGrid.cpp
@@ -153,7 +153,7 @@ void RenderTargetGrid::draw() {
 }
 
 double RenderTargetGrid::getAspectRatio(unsigned int vaoIndex) {
-	return drawWidth[vaoIndex] / drawHeight[vaoIndex];
+	return (double) drawWidth[vaoIndex] / (double) drawHeight[vaoIndex];
 }
 
 RenderTargetGrid::~RenderTargetGrid()
